Focus initial test

+++ src/main/java/com/revature/assignforce/service/FocusServiceImpl.java
@@ -1,7 +1,10 @@
 package com.revature.assignforce.service;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import javax.transaction.Transactional;
 
@@ -9,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.revature.assignforce.beans.Focus;
+import com.revature.assignforce.beans.SkillIdHolder;
 import com.revature.assignforce.repos.FocusRepository;
 
 @Transactional
@@ -35,12 +39,23 @@ public class FocusServiceImpl implements FocusService {
 
 	@Override
 	public Focus create(Focus b) {
+		Set<SkillIdHolder> ids = new HashSet<SkillIdHolder>();
+		ids.addAll(b.getSkills());
+		b.setSkills(null);
+		focusRepository.save(b);
+		b.setSkills(ids);
 		return focusRepository.save(b);
 	}
 
 	@Override
 	public void delete(int id) {
-		focusRepository.deleteById(id);
+		Optional<Focus> focus = focusRepository.findById(id);
+		if(focus.isPresent()) {
+			Focus f = focus.get();
+			f.setSkills(null);
+			focusRepository.save(f);
+			focusRepository.deleteById(id);
+		}
 	}
 
 }
